services/kira_til/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==1.10.13
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.9.3
